requirements.txt
Pillow==10.0.0
pyautogui==0.9.54
pynput==1.7.6